backend/requirements.txt
flask
flask-cors
ultralytics
transformers
torch
pillow
opencv-python-headless